@@ -5,225 +5,524 @@ package ir.co.yalda.formbuilder.utility
  */
 object Json {
     const val json = "{\n" +
-            "  \"MainFormCode\": 25,\n" +
-            "  \"MainFormVersion\": \"25.1\",\n" +
-            "  \"MainFormTitle\": \"مشاهده قبوض صادر شده\",\n" +
-            "  \"MainFormEnglishName\": \"MainFrmViewBill\",\n" +
-            "  \"MainFormSubmitUrl\": null,\n" +
-            "  \"MainFormOutputType\": \"None|Table|Message|Graph|PlainJson\",\n" +
-            "  \"FormsViewMode\": \"Tabular|GroupBox\",\n" +
-            "  \"Forms\": [\n" +
-            "    {\n" +
-            "      \"FormCode\": 12,\n" +
-            "      \"FormVersion\": \"12.1\",\n" +
-            "      \"FormTitle\": \"مشاهده قبوض صادر شده دوره اي\",\n" +
-            "      \"FormEnglishName\": \"FrmViewBillPerPeriod\",\n" +
-            "      \"SubmitUrl\": \"http://192.168.127.49/api/customer/viewBillPerPeriod\",\n" +
-            "      \"OutputType\": \"None|Table|Message|Graph|PlainJson\",\n" +
-            "      \"FormCondition\": \"(NumYear!=null || DropYear!=null) && (NumYear>=1390 && NumYear<=1397)\",\n" +
-            "      \"Elements\": [\n" +
+            "    \"Data\": [\n" +
             "        {\n" +
-            "          \"Type\": \"Numeric\",\n" +
-            "          \"Name\": \"NumYear\",\n" +
-            "          \"Title\": \"سال\",\n" +
-            "          \"IsUsedAsInputParam\": true,\n" +
-            "          \"DefaultValue\": 1396,\n" +
-            "          \"Validation\": {\n" +
-            "            \"IsRequired\": false\n" +
-            "          },\n" +
-            "          \"Style\": {\n" +
-            "            \"Dimension\": {\n" +
-            "              \"PhoneVerticalUnit\": \"0:6\",\n" +
-            "              \"PhoneHorizontalUnit\": \"0:3\",\n" +
-            "              \"TabletVerticalUnit\": \"0:3\",\n" +
-            "              \"TabletHorizontalUnit\": \"0:2\"\n" +
-            "            },\n" +
-            "            \"Type\": \"YaldaDefault|Material\",\n" +
-            "            \"BackgroundColor\": \"#abff44\",\n" +
-            "            \"FontColor\": \"black\",\n" +
-            "            \"Font\": \"Tahoma\",\n" +
-            "            \"FontSize\": \"12px\",\n" +
-            "            \"IconStyle\": \"fa-calendar\",\n" +
-            "            \"Height\": \"40px\",\n" +
-            "            \"TextAlign\": \"center\",\n" +
-            "            \"Direction\": \"LTR\"\n" +
-            "          }\n" +
+            "            \"VT\": \"7b547ed3314b1fe57d5ec740dd3fb524\",\n" +
+            "            \"List\": null,\n" +
+            "            \"Label\": \"Date Not Issued\",\n" +
+            "            \"DataType\": \"Date\",\n" +
+            "            \"SequenceNo\": 1,\n" +
+            "            \"IsMandatory\": 1,\n" +
+            "            \"StatusQueryId\": 65,\n" +
+            "            \"DataTypeSetting\": null\n" +
             "        },\n" +
             "        {\n" +
-            "          \"Type\": \"Dropdown\",\n" +
-            "          \"Name\": \"DropYear\",\n" +
-            "          \"Title\": \"سال\",\n" +
-            "          \"IsUsedAsInputParam\": true,\n" +
-            "          \"DefaultValue\": 1396,\n" +
-            "          \"List\": [\n" +
-            "            {\n" +
-            "              \"Text\": 1396,\n" +
-            "              \"Value\": 1396\n" +
-            "            },\n" +
-            "            {\n" +
-            "              \"Text\": 1395,\n" +
-            "              \"Value\": 1395\n" +
-            "            },\n" +
-            "            {\n" +
-            "              \"Text\": 1394,\n" +
-            "              \"Value\": 1394\n" +
+            "            \"VT\": \"30b4b5626505dd0655cf1f01eacf34a5\",\n" +
+            "            \"List\": null,\n" +
+            "            \"Label\": \"Not issued Image\",\n" +
+            "            \"DataType\": \"File\",\n" +
+            "            \"SequenceNo\": 2,\n" +
+            "            \"IsMandatory\": 1,\n" +
+            "            \"StatusQueryId\": 66,\n" +
+            "            \"DataTypeSetting\": {\n" +
+            "                \"File\": {\n" +
+            "                    \"MaxSize\": \"100\",\n" +
+            "                    \"MinSize\": \"10\",\n" +
+            "                    \"Exnetions\": \"jpg,png,gif\",\n" +
+            "                    \"CameraIsNeeded\": true\n" +
+            "                }\n" +
             "            }\n" +
-            "          ],\n" +
-            "          \"Validation\": {\n" +
-            "            \"IsRequired\": false\n" +
-            "          },\n" +
-            "          \"GetDataUrl\": \"http://192.168.127.49/api/pub/getPeriodYearByBillID\",\n" +
-            "          \"Style\": {\n" +
-            "            \"Dimension\": {\n" +
-            "              \"PhoneVerticalUnit\": \"0:6\",\n" +
-            "              \"PhoneHorizontalUnit\": \"0:3\",\n" +
-            "              \"TabletVerticalUnit\": \"0:3\",\n" +
-            "              \"TabletHorizontalUnit\": \"0:2\"\n" +
-            "            },\n" +
-            "            \"Type\": \"YaldaDefault|Material\",\n" +
-            "            \"BackgroundColor\": \"#abff44\",\n" +
-            "            \"FontColor\": \"black\",\n" +
-            "            \"Font\": \"Tahoma\",\n" +
-            "            \"FontSize\": \"12px\",\n" +
-            "            \"IconStyle\": \"fa-calendar\",\n" +
-            "            \"Height\": \"40px\",\n" +
-            "            \"TextAlign\": \"center\",\n" +
-            "            \"Direction\": \"LTR\"\n" +
-            "          }\n" +
             "        },\n" +
             "        {\n" +
-            "          \"Type\": \"Dropdown\",\n" +
-            "          \"Name\": \"DropPeriod\",\n" +
-            "          \"Title\": \"دوره\",\n" +
-            "          \"IsUsedAsInputParam\": true,\n" +
-            "          \"DefaultValue\": 6,\n" +
-            "          \"List\": [\n" +
-            "            {\n" +
-            "              \"Text\": \"دوره ششم\",\n" +
-            "              \"Value\": 6\n" +
-            "            },\n" +
-            "            {\n" +
-            "              \"Text\": \"دوره پنجم\",\n" +
-            "              \"Value\": 5\n" +
-            "            },\n" +
-            "            {\n" +
-            "              \"Text\": \"دوره چهارم\",\n" +
-            "              \"Value\": 4\n" +
-            "            },\n" +
-            "            {\n" +
-            "              \"Text\": \"دوره سوم\",\n" +
-            "              \"Value\": 3\n" +
-            "            },\n" +
-            "            {\n" +
-            "              \"Text\": \"دوره دوم\",\n" +
-            "              \"Value\": 2\n" +
-            "            },\n" +
-            "            {\n" +
-            "              \"Text\": \"دوره اول\",\n" +
-            "              \"Value\": 1\n" +
+            "            \"VT\": \"7f91e56cc201e95c73ef0b037cc3a872\",\n" +
+            "            \"List\": [\n" +
+            "                {\n" +
+            "                    \"VT\": \"d703e7565d26a79bcc328ecaaf140b02\",\n" +
+            "                    \"Code\": \"Citation\",\n" +
+            "                    \"ModuleName\": \"Transit\",\n" +
+            "                    \"SequenceNo\": 1,\n" +
+            "                    \"Description\": \"Citation\",\n" +
+            "                    \"GPSIsNeeded\": 0,\n" +
+            "                    \"CommentIsNeeded\": 0,\n" +
+            "                    \"CustomActionCode\": \"ReturnToOffice\",\n" +
+            "                    \"StatusQueryIssueId\": 76,\n" +
+            "                    \"IgnoredStatusQueryJson\": null\n" +
+            "                },\n" +
+            "                {\n" +
+            "                    \"VT\": \"5c9c396bbb4bba427d175b0f9d491c42\",\n" +
+            "                    \"Code\": null,\n" +
+            "                    \"ModuleName\": \"Transit\",\n" +
+            "                    \"SequenceNo\": 2,\n" +
+            "                    \"Description\": \"Jurisdiction\",\n" +
+            "                    \"GPSIsNeeded\": 0,\n" +
+            "                    \"CommentIsNeeded\": 0,\n" +
+            "                    \"CustomActionCode\": \"ReturnToOffice\",\n" +
+            "                    \"StatusQueryIssueId\": 78,\n" +
+            "                    \"IgnoredStatusQueryJson\": null\n" +
+            "                },\n" +
+            "                {\n" +
+            "                    \"VT\": \"5dd1036e411d2b7ca225e25f60a368d1\",\n" +
+            "                    \"Code\": null,\n" +
+            "                    \"ModuleName\": \"Transit\",\n" +
+            "                    \"SequenceNo\": 3,\n" +
+            "                    \"Description\": \"Not Signed\",\n" +
+            "                    \"GPSIsNeeded\": 0,\n" +
+            "                    \"CommentIsNeeded\": 0,\n" +
+            "                    \"CustomActionCode\": \"ReturnToOffice\",\n" +
+            "                    \"StatusQueryIssueId\": 84,\n" +
+            "                    \"IgnoredStatusQueryJson\": null\n" +
+            "                },\n" +
+            "                {\n" +
+            "                    \"VT\": \"6ed9a56ba1326cc7a8a928be59593509\",\n" +
+            "                    \"Code\": null,\n" +
+            "                    \"ModuleName\": \"Transit\",\n" +
+            "                    \"SequenceNo\": 4,\n" +
+            "                    \"Description\": \"No NCR Certificate\",\n" +
+            "                    \"GPSIsNeeded\": 0,\n" +
+            "                    \"CommentIsNeeded\": 0,\n" +
+            "                    \"CustomActionCode\": \"ReturnToOffice\",\n" +
+            "                    \"StatusQueryIssueId\": 81,\n" +
+            "                    \"IgnoredStatusQueryJson\": null\n" +
+            "                },\n" +
+            "                {\n" +
+            "                    \"VT\": \"c3d7d62cee3a3639b3ade16e237e9ad9\",\n" +
+            "                    \"Code\": null,\n" +
+            "                    \"ModuleName\": \"Transit\",\n" +
+            "                    \"SequenceNo\": 5,\n" +
+            "                    \"Description\": \"No Agreement\",\n" +
+            "                    \"GPSIsNeeded\": 0,\n" +
+            "                    \"CommentIsNeeded\": 0,\n" +
+            "                    \"CustomActionCode\": \"ReturnToOffice\",\n" +
+            "                    \"StatusQueryIssueId\": 79,\n" +
+            "                    \"IgnoredStatusQueryJson\": null\n" +
+            "                },\n" +
+            "                {\n" +
+            "                    \"VT\": \"c1d5611b7ec8a26eaf1d6a650bda1d75\",\n" +
+            "                    \"Code\": null,\n" +
+            "                    \"ModuleName\": \"Transit\",\n" +
+            "                    \"SequenceNo\": 6,\n" +
+            "                    \"Description\": \"No Annexures\",\n" +
+            "                    \"GPSIsNeeded\": 0,\n" +
+            "                    \"CommentIsNeeded\": 0,\n" +
+            "                    \"CustomActionCode\": \"ReturnToOffice\",\n" +
+            "                    \"StatusQueryIssueId\": 80,\n" +
+            "                    \"IgnoredStatusQueryJson\": null\n" +
+            "                },\n" +
+            "                {\n" +
+            "                    \"VT\": \"d94231dff5dc8f46c1a3c65c6c80cff5\",\n" +
+            "                    \"Code\": null,\n" +
+            "                    \"ModuleName\": \"Transit\",\n" +
+            "                    \"SequenceNo\": 7,\n" +
+            "                    \"Description\": \"No Place for Clerk to Sign\",\n" +
+            "                    \"GPSIsNeeded\": 0,\n" +
+            "                    \"CommentIsNeeded\": 0,\n" +
+            "                    \"CustomActionCode\": \"ReturnToOffice\",\n" +
+            "                    \"StatusQueryIssueId\": 82,\n" +
+            "                    \"IgnoredStatusQueryJson\": null\n" +
+            "                },\n" +
+            "                {\n" +
+            "                    \"VT\": \"f3a16bc4fb2465a783280e55601ce20d\",\n" +
+            "                    \"Code\": null,\n" +
+            "                    \"ModuleName\": \"Transit\",\n" +
+            "                    \"SequenceNo\": 8,\n" +
+            "                    \"Description\": \"Comply with NCR Act\",\n" +
+            "                    \"GPSIsNeeded\": 0,\n" +
+            "                    \"CommentIsNeeded\": 0,\n" +
+            "                    \"CustomActionCode\": \"ReturnToOffice\",\n" +
+            "                    \"StatusQueryIssueId\": 77,\n" +
+            "                    \"IgnoredStatusQueryJson\": null\n" +
+            "                },\n" +
+            "                {\n" +
+            "                    \"VT\": \"c4a3a195b05b6cd71579d72f73405f1b\",\n" +
+            "                    \"Code\": null,\n" +
+            "                    \"ModuleName\": \"Transit\",\n" +
+            "                    \"SequenceNo\": 9,\n" +
+            "                    \"Description\": \"No Track in Trace Report\",\n" +
+            "                    \"GPSIsNeeded\": 0,\n" +
+            "                    \"CommentIsNeeded\": 0,\n" +
+            "                    \"CustomActionCode\": \"ReturnToOffice\",\n" +
+            "                    \"StatusQueryIssueId\": 83,\n" +
+            "                    \"IgnoredStatusQueryJson\": null\n" +
+            "                },\n" +
+            "                {\n" +
+            "                    \"VT\": \"a670bd05e02e66bbd3d6380f930cab39\",\n" +
+            "                    \"Code\": null,\n" +
+            "                    \"ModuleName\": \"Transit\",\n" +
+            "                    \"SequenceNo\": 10,\n" +
+            "                    \"Description\": \"Other\",\n" +
+            "                    \"GPSIsNeeded\": 0,\n" +
+            "                    \"CommentIsNeeded\": 1,\n" +
+            "                    \"CustomActionCode\": \"ReturnToOffice\",\n" +
+            "                    \"StatusQueryIssueId\": 85,\n" +
+            "                    \"IgnoredStatusQueryJson\": null\n" +
+            "                },\n" +
+            "                {\n" +
+            "                    \"VT\": \"ae53b4b2baea76b17de5d64cb19aaaf9\",\n" +
+            "                    \"Code\": null,\n" +
+            "                    \"ModuleName\": \"Transit\",\n" +
+            "                    \"SequenceNo\": 11,\n" +
+            "                    \"Description\": \"No Clerk\",\n" +
+            "                    \"GPSIsNeeded\": 1,\n" +
+            "                    \"CommentIsNeeded\": 0,\n" +
+            "                    \"CustomActionCode\": \"ReturnToOffice\",\n" +
+            "                    \"StatusQueryIssueId\": 91,\n" +
+            "                    \"IgnoredStatusQueryJson\": [\n" +
+            "                        {\n" +
+            "                            \"StatusQueryId\": 66// this field should be hide and and not send\n" +
+            "                        }\n" +
+            "                    ]\n" +
+            "                },\n" +
+            "                {\n" +
+            "                    \"VT\": \"79250972f03493e86f9874d4c6cb4908\",\n" +
+            "                    \"Code\": \"WrongCourt\",\n" +
+            "                    \"ModuleName\": \"Jurisdiction\",\n" +
+            "                    \"SequenceNo\": 13,\n" +
+            "                    \"Description\": \"Wrong Court\",\n" +
+            "                    \"GPSIsNeeded\": 1,\n" +
+            "                    \"CommentIsNeeded\": 0,\n" +
+            "                    \"CustomActionCode\": \"WrongCourt\",//call get court //get sherif\n" +
+            "                    \"StatusQueryIssueId\": 86,\n" +
+            "                    \"IgnoredStatusQueryJson\": null\n" +
+            "                }\n" +
+            "            ],\n" +
+            "            \"Label\": \"Reason\",\n" +
+            "            \"DataType\": \"List\",\n" +
+            "            \"SequenceNo\": 3,\n" +
+            "            \"IsMandatory\": 1,\n" +
+            "            \"StatusQueryId\": 67,\n" +
+            "            \"DataTypeSetting\": null\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"VT\": \"d06292be726f2b1dcfe7f1c63e005961\",\n" +
+            "            \"List\": null,\n" +
+            "            \"Label\": \"PDF File\",\n" +
+            "            \"DataType\": \"File\",\n" +
+            "            \"SequenceNo\": 4,\n" +
+            "            \"IsMandatory\": 1,\n" +
+            "            \"StatusQueryId\": 68,\n" +
+            "            \"DataTypeSetting\": {\n" +
+            "                \"File\": {\n" +
+            "                    \"MaxSize\": \"10000000\",\n" +
+            "                    \"MinSize\": \"10\",\n" +
+            "                    \"Exnetions\": \"pdf\",\n" +
+            "                    \"CameraIsNeeded\": false\n" +
+            "                }\n" +
             "            }\n" +
-            "          ],\n" +
-            "          \"Validation\": {\n" +
-            "            \"IsRequired\": true\n" +
-            "          },\n" +
-            "          \"Style\": {\n" +
-            "            \"Dimension\": {\n" +
-            "              \"PhoneVerticalUnit\": \"0:6\",\n" +
-            "              \"PhoneHorizontalUnit\": \"3:3\",\n" +
-            "              \"TabletVerticalUnit\": \"3:3\",\n" +
-            "              \"TabletHorizontalUnit\": \"4:2\"\n" +
-            "            },\n" +
-            "            \"Type\": \"YaldaDefault|Material\",\n" +
-            "            \"BackgroundColor\": \"#abff44\",\n" +
-            "            \"FontColor\": \"black\",\n" +
-            "            \"Font\": \"Tahoma\",\n" +
-            "            \"FontSize\": \"12px\",\n" +
-            "            \"IconStyle\": \"fa-calendar\",\n" +
-            "            \"Height\": \"40px\",\n" +
-            "            \"TextAlign\": \"center\",\n" +
-            "            \"Direction\": \"LTR\"\n" +
-            "          }\n" +
             "        },\n" +
             "        {\n" +
-            "          \"Type\": \"Checkbox\",\n" +
-            "          \"Name\": \"ChkShowWithCanceled\",\n" +
-            "          \"Title\": \"آيا قبوض ابطال شده نيز نمايش داده شوند؟\",\n" +
-            "          \"IsUsedAsInputParam\": true,\n" +
-            "          \"DefaultValue\": false,\n" +
-            "          \"Validation\": null,\n" +
-            "          \"Style\": {\n" +
-            "            \"Dimension\": {\n" +
-            "              \"PhoneVerticalUnit\": \"0:3\",\n" +
-            "              \"PhoneHorizontalUnit\": \"2:3\",\n" +
-            "              \"TabletVerticalUnit\": \"0:3\",\n" +
-            "              \"TabletHorizontalUnit\": \"3:3\"\n" +
-            "            },\n" +
-            "            \"Type\": \"YaldaDefault|Material\",\n" +
-            "            \"BackgroundColor\": \"#abff44\",\n" +
-            "            \"FontColor\": \"black\",\n" +
-            "            \"Font\": \"Tahoma\",\n" +
-            "            \"FontSize\": \"12px\",\n" +
-            "            \"IconStyle\": \"fa-calendar\",\n" +
-            "            \"Height\": \"40px\",\n" +
-            "            \"TextAlign\": \"center\",\n" +
-            "            \"Direction\": \"LTR\"\n" +
-            "          }\n" +
-            "        },\n" +
-            "        {\n" +
-            "          \"Type\": \"TextArea\",\n" +
-            "          \"Name\": \"TxtDescription\",\n" +
-            "          \"Title\": \"توضيحات\",\n" +
-            "          \"IsUsedAsInputParam\": true,\n" +
-            "          \"DefaultValue\": null,\n" +
-            "          \"Validation\": {\n" +
-            "            \"IsRequired\": false\n" +
-            "          },\n" +
-            "          \"Style\": {\n" +
-            "            \"Dimension\": {\n" +
-            "              \"PhoneVerticalUnit\": \"0:6\",\n" +
-            "              \"PhoneHorizontalUnit\": \"0:6\",\n" +
-            "              \"TabletVerticalUnit\": \"0:6\",\n" +
-            "              \"TabletHorizontalUnit\": \"0:6\"\n" +
-            "            },\n" +
-            "            \"Type\": \"YaldaDefault|Material\",\n" +
-            "            \"BackgroundColor\": \"#abff44\",\n" +
-            "            \"FontColor\": \"black\",\n" +
-            "            \"Font\": \"Tahoma\",\n" +
-            "            \"FontSize\": \"12px\",\n" +
-            "            \"IconStyle\": \"fa-calendar\",\n" +
-            "            \"Height\": \"40px\",\n" +
-            "            \"TextAlign\": \"center\",\n" +
-            "            \"Direction\": \"LTR\"\n" +
-            "          }\n" +
-            "        },\n" +
-            "        {\n" +
-            "          \"Type\": \"Submit\",\n" +
-            "          \"Name\": \"BtnViewBillPerPeriod\",\n" +
-            "          \"Title\": \"مشاهده\",\n" +
-            "          \"IsUsedAsInputParam\": false,\n" +
-            "          \"SubmitUrl\": null,\n" +
-            "          \"OutputType\": \"None|Message|PlainJson\",\n" +
-            "          \"Validation\": null,\n" +
-            "          \"Style\": {\n" +
-            "            \"Dimension\": {\n" +
-            "              \"PhoneVerticalUnit\": \"0:6\",\n" +
-            "              \"PhoneHorizontalUnit\": \"0:3\",\n" +
-            "              \"TabletVerticalUnit\": \"0:3\",\n" +
-            "              \"TabletHorizontalUnit\": \"0:2\"\n" +
-            "            },\n" +
-            "            \"Type\": \"YaldaDefault|Material\",\n" +
-            "            \"BackgroundColor\": \"#abff44\",\n" +
-            "            \"FontColor\": \"black\",\n" +
-            "            \"Font\": \"Tahoma\",\n" +
-            "            \"FontSize\": \"12px\",\n" +
-            "            \"IconStyle\": \"fa-calendar\",\n" +
-            "            \"Height\": \"40px\",\n" +
-            "            \"TextAlign\": \"center\",\n" +
-            "            \"Direction\": \"LTR\"\n" +
-            "          }\n" +
+            "            \"VT\": \"81df19f3eeb06a2f0f29d7683768af63\",\n" +
+            "            \"List\": null,\n" +
+            "            \"Label\": \"Person Name\",\n" +
+            "            \"DataType\": \"String\",\n" +
+            "            \"SequenceNo\": 5,\n" +
+            "            \"IsMandatory\": 1,\n" +
+            "            \"StatusQueryId\": 69,\n" +
+            "            \"DataTypeSetting\": null\n" +
             "        }\n" +
-            "      ]\n" +
+            "    ],\n" +
+            "    \"ErrorHandling\": {\n" +
+            "        \"ErrorCode\": 0,\n" +
+            "        \"ErrorMustBeSeenByUser\": false,\n" +
+            "        \"IsSuccessful\": true,\n" +
+            "        \"ErrorLevel\": \"Info\"\n" +
             "    }\n" +
-            "  ]\n" +
             "}"
+
+    const val courts = "[\n" +
+            "        {\n" +
+            "            \"CourtId\": 1,\n" +
+            "            \"CourtName\": \"ABERDEEN MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 3,\n" +
+            "            \"CourtName\": \"ADELAIDE MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 4,\n" +
+            "            \"CourtName\": \"ALEXANDRIA MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 5,\n" +
+            "            \"CourtName\": \"ALICE MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 7,\n" +
+            "            \"CourtName\": \"ALIWAL NORTH MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 9,\n" +
+            "            \"CourtName\": \"BARKLY EAST MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 10,\n" +
+            "            \"CourtName\": \"BEDFORD MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 13,\n" +
+            "            \"CourtName\": \"BIZANA MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 15,\n" +
+            "            \"CourtName\": \"BURGERSDORP MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 16,\n" +
+            "            \"CourtName\": \"BUTTERWORTH MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 17,\n" +
+            "            \"CourtName\": \"CALA MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 18,\n" +
+            "            \"CourtName\": \"CATHCART MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 19,\n" +
+            "            \"CourtName\": \"CENTANE MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 20,\n" +
+            "            \"CourtName\": \"COFIMVABA MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 22,\n" +
+            "            \"CourtName\": \"CRADOCK MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 25,\n" +
+            "            \"CourtName\": \"DORDRECHT MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 26,\n" +
+            "            \"CourtName\": \"DUTYWA MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 27,\n" +
+            "            \"CourtName\": \"EAST LONDON MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 28,\n" +
+            "            \"CourtName\": \"ELLIOT MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 29,\n" +
+            "            \"CourtName\": \"ELLIOTDALE MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 30,\n" +
+            "            \"CourtName\": \"EZIBELENI MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 31,\n" +
+            "            \"CourtName\": \"FLAGSTAFF MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"CourtId\": 32,\n" +
+            "            \"CourtName\": \"FORT BEAUFORT MAGISTRATE COURT\",\n" +
+            "            \"CourtTypeId\": 4,\n" +
+            "            \"CourtType\": \"MAGISTRATE COURT\"\n" +
+            "        }" +
+            "]"
+
+    const val sheriffs = "[\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 422,\n" +
+            "            \"CourtId\": 1,\n" +
+            "            \"CourtName\": \"ABERDEEN MAGISTRATE COURT\",\n" +
+            "            \"SheriffOfficeId\": 156,\n" +
+            "            \"SheriffAreaName\": \"ABERDEEN SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 758,\n" +
+            "            \"CourtId\": 3,\n" +
+            "            \"CourtName\": \"ADELAIDE MAGISTRATE COURT\",\n" +
+            "            \"SheriffOfficeId\": 173,\n" +
+            "            \"SheriffAreaName\": \"ADELAIDE SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 991,\n" +
+            "            \"CourtId\": 209,\n" +
+            "            \"CourtName\": \"ALEXANDRA MAGISTRATE COURT\",\n" +
+            "            \"SheriffOfficeId\": 502,\n" +
+            "            \"SheriffAreaName\": \"ALEXANDRA SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 354,\n" +
+            "            \"CourtId\": 4,\n" +
+            "            \"CourtName\": \"ALEXANDRIA MAGISTRATE COURT\",\n" +
+            "            \"SheriffOfficeId\": 167,\n" +
+            "            \"SheriffAreaName\": \"ALEXANDRIA SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 768,\n" +
+            "            \"CourtId\": 5,\n" +
+            "            \"CourtName\": \"ALICE MAGISTRATE COURT\",\n" +
+            "            \"SheriffOfficeId\": 171,\n" +
+            "            \"SheriffAreaName\": \"ALICE SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 262,\n" +
+            "            \"CourtId\": 7,\n" +
+            "            \"CourtName\": \"ALIWAL NORTH MAGISTRATE COURT\",\n" +
+            "            \"SheriffOfficeId\": 223,\n" +
+            "            \"SheriffAreaName\": \"ALIWAL NORTH SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 462,\n" +
+            "            \"CourtId\": 502,\n" +
+            "            \"CourtName\": \"AMERSFOORT MAGISTRATE COURT\",\n" +
+            "            \"SheriffOfficeId\": 88,\n" +
+            "            \"SheriffAreaName\": \"AMERSFOORT SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 963,\n" +
+            "            \"CourtId\": 503,\n" +
+            "            \"CourtName\": \"AMSTERDAM MAGISTRATE COURT\",\n" +
+            "            \"SheriffOfficeId\": 699,\n" +
+            "            \"SheriffAreaName\": \"AMSTERDAM SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 957,\n" +
+            "            \"CourtId\": 597,\n" +
+            "            \"CourtName\": \"ATAMELANG MAGISTRATE COURT\",\n" +
+            "            \"SheriffOfficeId\": 675,\n" +
+            "            \"SheriffAreaName\": \"ATAMELANG SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 879,\n" +
+            "            \"CourtId\": 725,\n" +
+            "            \"CourtName\": \"ATLANTIS MAGISTRATE COURT\",\n" +
+            "            \"SheriffOfficeId\": 414,\n" +
+            "            \"SheriffAreaName\": \"ATLANTIS SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 903,\n" +
+            "            \"CourtId\": 210,\n" +
+            "            \"CourtName\": \"ATTERIDGEVILLE MAGISTRATE COURT\",\n" +
+            "            \"SheriffOfficeId\": 643,\n" +
+            "            \"SheriffAreaName\": \"ATTERIDGEVILLE SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 434,\n" +
+            "            \"CourtId\": 271,\n" +
+            "            \"CourtName\": \"BABANANGO MAGISTRATE COURT\",\n" +
+            "            \"SheriffOfficeId\": 112,\n" +
+            "            \"SheriffAreaName\": \"BABANANGO SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 786,\n" +
+            "            \"CourtId\": 504,\n" +
+            "            \"CourtName\": \"BALFOUR [MP] MAGISTRATE COURT\",\n" +
+            "            \"SheriffOfficeId\": 22,\n" +
+            "            \"SheriffAreaName\": \"BALFOUR SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 2366,\n" +
+            "            \"CourtId\": 903,\n" +
+            "            \"CourtName\": \"BALFOUR REGIONAL COURT\",\n" +
+            "            \"SheriffOfficeId\": 22,\n" +
+            "            \"SheriffAreaName\": \"BALFOUR SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 464,\n" +
+            "            \"CourtId\": 505,\n" +
+            "            \"CourtName\": \"BARBERTON MAGISTRATE COURT\",\n" +
+            "            \"SheriffOfficeId\": 78,\n" +
+            "            \"SheriffAreaName\": \"BARBERTON SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 2397,\n" +
+            "            \"CourtId\": 916,\n" +
+            "            \"CourtName\": \"BARBERTON REGIONAL COURT\",\n" +
+            "            \"SheriffOfficeId\": 78,\n" +
+            "            \"SheriffAreaName\": \"BARBERTON SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 591,\n" +
+            "            \"CourtId\": 9,\n" +
+            "            \"CourtName\": \"BARKLY EAST MAGISTRATE COURT\",\n" +
+            "            \"SheriffOfficeId\": 225,\n" +
+            "            \"SheriffAreaName\": \"BARKLY EAST SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 1217,\n" +
+            "            \"CourtId\": 848,\n" +
+            "            \"CourtName\": \"BARKLY EAST REGIONAL COURT\",\n" +
+            "            \"SheriffOfficeId\": 223,\n" +
+            "            \"SheriffAreaName\": \"ALIWAL NORTH SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 1221,\n" +
+            "            \"CourtId\": 848,\n" +
+            "            \"CourtName\": \"BARKLY EAST REGIONAL COURT\",\n" +
+            "            \"SheriffOfficeId\": 225,\n" +
+            "            \"SheriffAreaName\": \"BARKLY EAST SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 1216,\n" +
+            "            \"CourtId\": 848,\n" +
+            "            \"CourtName\": \"BARKLY EAST REGIONAL COURT\",\n" +
+            "            \"SheriffOfficeId\": 204,\n" +
+            "            \"SheriffAreaName\": \"BURGERSDORP SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 1218,\n" +
+            "            \"CourtId\": 848,\n" +
+            "            \"CourtName\": \"BARKLY EAST REGIONAL COURT\",\n" +
+            "            \"SheriffOfficeId\": 222,\n" +
+            "            \"SheriffAreaName\": \"DORDRECHT SHERIFF\"\n" +
+            "        },\n" +
+            "        {\n" +
+            "            \"SheriffLinkId\": 1223,\n" +
+            "            \"CourtId\": 848,\n" +
+            "            \"CourtName\": \"BARKLY EAST REGIONAL COURT\",\n" +
+            "            \"SheriffOfficeId\": 288,\n" +
+            "            \"SheriffAreaName\": \"HERSCHEL SHERIFF\"\n" +
+            "        }" +
+            "]"
 }
